// src/runtime/server/types/storeApiTypes.d.ts
type GenericRecord = never | null | string | string[] | number | { [key: string]: GenericRecord };
export type components = {
  schemas: Schemas;
};
export type Schemas = {
// ...
  Media: {
    alt?: string;
    /** @enum {string} */
    apiAlias: 'media';
    config?: GenericRecord;
    /** Format: date-time */
    createdAt: string;
    customFields?: GenericRecord;
    fileExtension: string;
    fileName: string;
    /** Format: int64 */
    fileSize?: number;
    /** Runtime field, cannot be used as part of the criteria. */
    hasFile: boolean;
    id: string;
    metaData?: {
      /** Format: int64 */
      height?: number;
      /** Format: int64 */
      width?: number;
    };
    mimeType?: string;
    path: string;
    private: boolean;
    thumbnails?: components['schemas']['MediaThumbnail'][];
    title?: string;
    translated?: {
      alt?: string;
      apiAlias?: string;
      fileExtension?: string;
      fileName?: string;
      mimeType?: string;
      path?: string;
      title?: string;
      uploadedAt?: string;
      url?: string;
    };
    /** Format: date-time */
    updatedAt?: string;
    /** Format: date-time */
    uploadedAt?: string;
    /** Runtime field, cannot be used as part of the criteria. */
    url: string;
  };
// ...
  MediaThumbnail: {
    /** Format: date-time */
    createdAt: string;
    customFields?: GenericRecord;
    /** Format: int64 */
    height: number;
    id: string;
    mediaId: string;
    path?: string;
    /** Format: date-time */
    updatedAt?: string;
    /** Runtime field, cannot be used as part of the criteria. */
    url: string;
    /** Format: int64 */
    width: number;
  };
// ...
};